feat: new import system for Style
